pyproject.toml
[project]
name = "k-i-no"
version = "0.5"
description = "Having fun with movie data!"
dependencies = [
    "matplotlib>=3.8.4",
    "pandas>=2.2.2",
    "xlrd>=2.0.1",
    "python-dotenv>=1.0.1",
    "tqdm>=4.66.4",
    "scikit-learn>=1.4.2",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",
    "odfpy>=1.4.1",
    "openpyxl>=3.1.2",
    "flask>=3.0.3",
]
readme = "README.md"
requires-python = ">= 3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.0.285"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]
